Integrate self-update into agent exe
